app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "App todolist"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask==3.1.2",
    "flask-sqlalchemy==3.1.1",
    "gunicorn==23.0.0",
    "psycopg2-binary==2.9.10",
]

[project.optional-dependencies]
dev = [
    "ruff==0.14.3",
    "pytest==8.4.2",
    "pytest-flask==1.3.0",
    "pytest-cov==7.0.0",
    "pytest-html==4.1.1",
]

[tool.ruff]
# Configuration minimale pour Ruff
line-length = 100
target-version = "py313"

[tool.ruff.lint]
# Ignorer E402 (module level import not at top of file) dans les tests
per-file-ignores = { "tests/*.py" = ["E402"] }

[tool.pytest.ini_options]
# Configuration pour les tests
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py", "units.py", "integration.py", "regression.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Capture des logs pour les rapports HTML
log_cli = true
log_cli_level = "INFO"

# Markers pour organiser les tests
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "regression: Regression tests",
]

# Coverage
[tool.coverage.run]
source = ["app"]
omit = [
    "*/venv/*",
    "*/env/*",
    "*/__pycache__/*",
    "*/tests/*",
]

# Coverage report
[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
]
fail_under = 80

[tool.coverage.html]
directory = "coverage-html"
title = "Tests Coverage Report"